modified according to the comments recieved
